Eliminated player rules
Added 1.10A - Players section and rule to clarify the role of a Player after they are eliminated from a game they are participating in. This goes in line with the common-sense of “the dead don’t speak”, and also serves as a way to remove the potential for abuse of an eliminated player providing “outside assistance” to the other players in the pod in such a way to influence them towards a draw.